# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS ESTIMATION BACKEND DOCUMENTATION",
    "description": "Our story started in 2015 when there was a need of Construction database for helping bidders to bid accurately with reliable information's and estimates in the Rwanda construction industry. Continuous determination, hard work and passion have been constant themes throughout our history and present. As Rwanda Construction industry becomes increasingly important on the global stage, it's our believe that, with our own innovative mindset, we can make the construction industry a better one for national development. So we challenge ourselves to evolve, innovate and experiment to make work easier and faster for everyone as the leading local software brand in the Rwanda construction industry. We are currently the quantities and estimations software dealers which are made in Africa its self initially from Ghana and same technology has been customized to fit Rwanda construction industry and we are super proud to start the race with national development in minds. Today we are privileged to connect with thousands of Engineers and other construction related professionals who do daily estimates or quantities in the construction industry.",
    "version": "1.0.0",
    "contact": {
      "name": "DS ESTIMATION",
      "url": "http://dsestimation.com",
      "email": "dsestimation@gmail.com"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "AuthToken": {
      "type": "apiKey",
      "name": "auth-token",
      "in": "header",
      "description": "The token for authentication"
    }
  },
  "security": [
    {
      "AuthToken": []
    }
  ],
  "paths": {
...
    "/owner/unit/create": {
      "post": {
        "tags": ["Default Unit Routes"],
        "summary": "Create a unit",
        "description": "Create a unit",
        "operationId": "createUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Unit object that needs to be created unit",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unit created successfully"
          },
          "400": {
            "description": "Some unit field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/unit/{id}/update": {
      "patch": {
        "tags": ["Default Unit Routes"],
        "summary": "Update a unit",
        "description": "Update a unit",
        "operationId": "updateUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the unit to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Unit object that needs to be updated unit, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unit updated successfully"
          },
          "400": {
            "description": "Some unit field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Unit not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/unit/{id}/delete": {
      "delete": {
        "tags": ["Default Unit Routes"],
        "summary": "Delete a unit",
        "description": "Delete a unit",
        "operationId": "deleteUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the unit to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Unit deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Unit not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/unit/all": {
      "get": {
        "tags": ["Client Unit Routes"],
        "summary": "Get all units",
        "description": "Get all units",
        "operationId": "getAllUnits",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Units Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/user/unit/create": {
      "post": {
        "tags": ["Client Unit Routes"],
        "summary": "Create a unit",
        "description": "Create a unit",
        "operationId": "createUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Unit object that needs to be created unit",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unit created successfully"
          },
          "400": {
            "description": "Some unit field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/unit/{id}/update": {
      "patch": {
        "tags": ["Client Unit Routes"],
        "summary": "Update a unit",
        "description": "Update a unit",
        "operationId": "updateUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the unit to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Unit object that needs to be updated unit, you can update one or more fields",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unit updated successfully"
          },
          "400": {
            "description": "Some unit field is required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Unit not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user/unit/{id}/delete": {
      "delete": {
        "tags": ["Client Unit Routes"],
        "summary": "Delete a unit",
        "description": "Delete a unit",
        "operationId": "deleteUnit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of the unit to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Unit deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Unit not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/global/unit/all": {
      "get": {
        "tags": ["Global Routes"],
        "summary": "Get all units from our database",
        "description": "Get all units from our database",
        "operationId": "getAllUnits",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Units Retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Unit": {
      "type": "object",
      "properties": {
        "symbol": {
          "type": "string"
        },
        "meaning": {
          "type": "string"
        }
      }
    },
...
  }
}